Diagrams.SVG.Arguments

Attribute Parsing of classes of attributes

coreAttributes :: AttrParser CoreAttributes

conditionalProcessingAttributes :: AttrParser ConditionalProcessingAttributes

documentEventAttributes :: AttrParser DocumentEventAttributes

graphicalEventAttributes :: AttrParser GraphicalEventAttributes

presentationAttributes :: AttrParser PresentationAttributes

filterPrimitiveAttributes :: AttrParser FilterPrimitiveAttributes

xlinkAttributes :: AttrParser XlinkAttributes

xmlnsNameSpaces :: AttrParser NameSpaces

Attributes for basic structure elements

svgAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, NameSpaces, Maybe Text)

Attributes for <svg>, see http://www.w3.org/TR/SVG/struct.html#SVGElement

gAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <g> and <defs>, see http://www.w3.org/TR/SVG/struct.html#GElement

sAttrs :: AttrParser (CoreAttributes, Maybe Text, Maybe Text, Maybe Text)

Attributes for <g> and <defs>, see http://www.w3.org/TR/SVG/struct.html#GElement

descAttrs :: AttrParser (CoreAttributes, Maybe Text, Maybe Text)

Attributes for <desc>, see http://www.w3.org/TR/SVG/struct.html#DescriptionAndTitleElements

symbolAttrs :: AttrParser (CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <symbol>, see http://www.w3.org/TR/SVG/struct.html#SymbolElement

useAttrs :: AttrParser (CoreAttributes, ConditionalProcessingAttributes, GraphicalEventAttributes, PresentationAttributes, XlinkAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <use>, see http://www.w3.org/TR/SVG/struct.html#UseElement

switchAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <switch>, see http://www.w3.org/TR/SVG/struct.html#SwitchElement

Attributes for basic shape elements

rectAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <rect>, see http://www.w3.org/TR/SVG11/shapes.html#RectElement

circleAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <circle>, see http://www.w3.org/TR/SVG11/shapes.html#CircleElement

ellipseAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <ellipse>, see http://www.w3.org/TR/SVG11/shapes.html#EllipseElement

lineAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <line>, see http://www.w3.org/TR/SVG11/shapes.html#LineElement

polygonAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <polygon>, see http://www.w3.org/TR/SVG11/shapes.html#PolygonElement

pathAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <path>, see http://www.w3.org/TR/SVG11/paths.html#PathElement

Other Attributes

clipPathAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <clipPath>, see http://www.w3.org/TR/SVG/masking.html#ClipPathElement

patternAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <pattern>, see http://www.w3.org/TR/SVG/pservers.html#PatternElement

imageAttrs :: AttrParser (CoreAttributes, ConditionalProcessingAttributes, GraphicalEventAttributes, XlinkAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <image>, see http://www.w3.org/TR/SVG/struct.html#ImageElement

filterAttrs :: AttrParser (CoreAttributes, PresentationAttributes, XlinkAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <filter>, see http://www.w3.org/TR/SVG/filters.html#FilterElement

linearGradAttrs :: AttrParser (CoreAttributes, PresentationAttributes, XlinkAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

radialGradAttrs :: AttrParser (CoreAttributes, PresentationAttributes, XlinkAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

setAttrs :: AttrParser (CoreAttributes, PresentationAttributes, XlinkAttributes)

stopAttrs :: AttrParser (CoreAttributes, PresentationAttributes, XlinkAttributes, Maybe Text, Maybe Text, Maybe Text)

textAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <text>, see http://www.w3.org/TR/SVG/text.html#TextElement

tspanAttrs :: AttrParser (ConditionalProcessingAttributes, CoreAttributes, GraphicalEventAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

namedViewAttrs :: AttrParser (Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

perspectiveAttrs :: AttrParser (Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Font Attributes

fontAttrs :: AttrParser (CoreAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <font>

fontFaceAttrs :: AttrParser (CoreAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Attributes for <font-face>

glyphAttrs :: AttrParser (CoreAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

missingGlyphAttrs :: AttrParser (CoreAttributes, PresentationAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

kernAttrs :: AttrParser (CoreAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

Filter Effect Attributes

feBlendAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feColorMatrixAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feComponentTransferAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text)

feCompositeAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feConvolveMatrixAttrs :: AttrParser (CoreAttributes, Maybe Text, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feDiffuseLightingAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feDisplacementMapAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feFloodAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text)

feGaussianBlurAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feImageAttrs :: AttrParser (CoreAttributes, Maybe Text, FilterPrimitiveAttributes, XlinkAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feMergeAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text)

feMorphologyAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feOffsetAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feSpecularLightingAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)

feTileAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text)

feTurbulenceAttrs :: AttrParser (CoreAttributes, PresentationAttributes, FilterPrimitiveAttributes, Maybe Text, Maybe Text, Maybe Text, Maybe Text, Maybe Text)