Diagrams.SVG.Attributes

Description

Documentation

initialStyles :: (HasStyle c, Typeable (N c), Floating (N c), Ord (N c), V c ~ V2) => c -> c

Inital styles, see: http://www.w3.org/TR/SVG/painting.html#FillProperty

Classes of attributes

data CoreAttributes

Constructors

CA

Fields

id1 :: Maybe Text

xmlbase :: Maybe Text

xmllang :: Maybe Text

xmlspace :: Maybe Text

data ConditionalProcessingAttributes

Constructors

CPA

Fields

requiredFeatures :: Maybe Text

requiredExtensions :: Maybe Text

systemLanguage :: Maybe Text

data DocumentEventAttributes

Constructors

DEA

Fields

onunload :: Maybe Text

onabort :: Maybe Text

onerror :: Maybe Text

onresize :: Maybe Text

onscroll :: Maybe Text

onzoom :: Maybe Text

data GraphicalEventAttributes

Constructors

GEA

Fields

onfocusin :: Maybe Text

onfocusout :: Maybe Text

onactivate :: Maybe Text

onclick :: Maybe Text

onmousedown :: Maybe Text

onmouseup :: Maybe Text

onmouseover :: Maybe Text

onmousemove :: Maybe Text

onmouseout :: Maybe Text

onload :: Maybe Text

data XlinkAttributes

Constructors

XLA

Fields

xlinkHref :: Maybe Text

xlinkShow :: Maybe Text

xlinkActuate :: Maybe Text

xlinkType :: Maybe Text

xlinkRole :: Maybe Text

xlinkArcrole :: Maybe Text

xlinkTitle :: Maybe Text

data FilterPrimitiveAttributes

Constructors

FPA

Fields

x :: Maybe Text

y :: Maybe Text

width :: Maybe Text

height :: Maybe Text

result :: Maybe Text

data NameSpaces

Constructors

NSP

Fields

xlink :: Maybe Text

dc :: Maybe Text

cc :: Maybe Text

rdf :: Maybe Text

svg :: Maybe Text

sodipodi :: Maybe Text

inkscape :: Maybe Text

General Parsing Functions

separatedBy :: Parser Text a -> Text -> Parser Text [a]

Parsing content separated by something, e.g. ";" like in: "a;b;c;d;" or "a;b;c;d"

parseOne :: Parser Text b -> Text -> Parser Text b

parseOne' :: Parser Text b -> Parser Text b

compose :: [a -> a] -> a -> a

See http://www.haskell.org/haskellwiki/Compose

parseDouble :: RealFloat n => Text -> n

parseToDouble :: RealFloat n => Maybe Text -> Maybe n

parsePoints :: RealFloat n => Text -> [(n, n)]

parseTempl :: Parser a -> Maybe Text -> Maybe a

a template that deals with the common parser errors

parseIRI :: Parser Text (Text, Text)

Transformations

applyTr :: (RealFloat (N a), Transformable a, V a ~ V2) => [Transform (N a)] -> a -> a

parseTr :: RealFloat n => Maybe Text -> [Transform n]

Parsing the style attribute

applyStyleSVG :: (HasStyle a, Typeable (N a), RealFloat (N a), V a ~ V2) => (t -> [SVGStyle (N a) Double]) -> t -> a -> a

parseStyles :: (RealFloat n, RealFrac a, Floating a) => Maybe Text -> (HashMap Text (Tag b n), CSSMap, HashMap Text (Gr n)) -> [SVGStyle n a]

parseLengths :: Fractional n => Parser Text [LenPercent n]

parseViewBox :: RealFloat n => Maybe Text -> Maybe Text -> Maybe Text -> Maybe (ViewBox n)

Example: viewBox="0 0 100 30" Viewboxes establish a new viewport. Percentages (e.g. x="50%") only make sense with a viewport.

parsePA :: (RealFloat n, RealFloat a, Read a) => PresentationAttributes -> HashMaps b n -> [SVGStyle n a]

cssStylesFromMap :: (RealFloat n, RealFrac a, Floating a) => (HashMap Text (Tag b n), HashMap Text [(Text, Text)], HashMap Text (Gr n)) -> Text -> Maybe Text -> Maybe Text -> [SVGStyle n a]

This function is called on every tag and returns a list of style-attributes to apply (if there is a rule that matches) TO DO: CSS2 + CSS3 selectors cssStylesFromMap :: (Read a, RealFloat a, RealFloat n) => HashMaps b n -> Text -> Maybe Text -> Maybe Text -> [(SVGStyle n a)]

fragment :: Maybe Text -> Maybe Text

p :: RealFloat n => (n, n) -> n -> Maybe Text -> n

Given a minimum and maximum value of a viewbox (x or y-direction) and a maybe a Text value Parse this Text value as a length (with a unit) or a percentage relative to the viewbox (minx,maxx) If parsers fails return def

parseSpread :: Maybe Text -> SpreadMethod

Example: spreadMethod="pad"

Parsing Colors

Parsing preserve aspect ratio

parsePreserveAR :: Maybe Text -> Maybe PreserveAR

data PreserveAR

Constructors

PAR AlignSVG MeetOrSlice
see http://www.w3.org/TR/SVG11/coords.html#PreserveAspectRatioAttribute

data AlignSVG

Constructors

AlignXY Place Place
alignment in x and y direction

type Place

Arguments

= Double
A value between 0 and 1, where 0 is the minimal value and 1 the maximal value

data MeetOrSlice

Constructors

Meet

Slice

data SVGStyle n a

Constructors

Fill (AlphaColour a)

FillTex (Texture n)

FillOpacity Double

FillRule FR

Opacity Double

Stroke (AlphaColour a)

StrokeTex (Texture n)

StrokeWidth (LenPercent n)

StrokeLineCap LineCap

StrokeLineJoin LineJoin

StrokeMiterLimit n

StrokeDasharray [LenPercent n]

StrokeOpacity Double

FontFamily String

FontStyle FStyle

FontVariant FVariant

FontWeight FWeight

FontStretch FStretch

FontSize (LenPercent n)

ClipPath (Path V2 n)

EmptyStyle

data PresentationAttributes

Constructors

PA

Fields

alignmentBaseline :: Maybe Text

baselineShift :: Maybe Text

clip :: Maybe Text

clipPath :: Maybe Text

clipRule :: Maybe Text

color :: Maybe Text

colorInterpolation :: Maybe Text

colorInterpolationFilters :: Maybe Text

colorProfile :: Maybe Text

colorRendering :: Maybe Text

cursor :: Maybe Text

direction :: Maybe Text

display :: Maybe Text

dominantBaseline :: Maybe Text

enableBackground :: Maybe Text

fill :: Maybe Text

fillOpacity :: Maybe Text

fillRuleSVG :: Maybe Text

filter :: Maybe Text

floodColor :: Maybe Text

floodOpacity :: Maybe Text

fontFamily :: Maybe Text

fntSize :: Maybe Text

fontSizeAdjust :: Maybe Text

fontStretch :: Maybe Text

fontStyle :: Maybe Text

fontVariant :: Maybe Text

fontWeight :: Maybe Text

glyphOrientationHorizontal :: Maybe Text

glyphOrientationVertical :: Maybe Text

imageRendering :: Maybe Text

kerning :: Maybe Text

letterSpacing :: Maybe Text

lightingColor :: Maybe Text

markerEnd :: Maybe Text

markerMid :: Maybe Text

markerStart :: Maybe Text

mask :: Maybe Text

opacity :: Maybe Text

overflow :: Maybe Text

pointerEvents :: Maybe Text

shapeRendering :: Maybe Text

stopColor :: Maybe Text

stopOpacity :: Maybe Text

strokeSVG :: Maybe Text

strokeDasharray :: Maybe Text

strokeDashoffset :: Maybe Text

strokeLinecap :: Maybe Text

strokeLinejoin :: Maybe Text

strokeMiterlimit :: Maybe Text

strokeOpacity :: Maybe Text

strokeWidth :: Maybe Text

textAnchor :: Maybe Text

textDecoration :: Maybe Text

textRendering :: Maybe Text

unicodeBidi :: Maybe Text

visibility :: Maybe Text

wordSpacing :: Maybe Text

writingMode :: Maybe Text