Diagrams.TwoD.Input

Description

Importing external images into diagrams.

Documentation

loadImageEmbedded :: (InputConstraints b n, Renderable (Text n) b, Read n, n ~ Place) => String -> IO (Either String (QDiagram b V2 n Any))

Load 2d formats given by a filepath and embed them

loadImageExternal :: (InputConstraints b n, Renderable (DImage n External) b) => FilePath -> IO (Either String (QDiagram b V2 n Any))

Load 2d formats given by a filepath and make a reference